Refactor dashboard and buttons such that admins and users are separated completely
