+++ frontend/src/App.tsx
@@ -3,7 +3,6 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import QuestionPage from "./pages/QuestionPage";
 import LandingPage from "./pages/LandingPage";
 import DashboardPage from "./pages/DashboardPage";
-import DashboardPageForUsers from "./pages/DashboardForUsersPage";
 import LoginPage from "./pages/LoginPage";
 import RegisterPage from "./pages/RegisterPage";
 import ProfilePage from "./pages/ProfilePage";
@@ -18,10 +17,6 @@ const App: React.FC = () => {
           <Route path="/" element={<LandingPage />} />
           <Route path="/question/:title" element={<QuestionPage />} />
           <Route path="/dashboard" element={<DashboardPage />} />
-          <Route
-            path="/dashboardForUsers"
-            element={<DashboardPageForUsers />}
-          />
           <Route path="/login" element={<LoginPage />} />
           <Route path="/register" element={<RegisterPage />} />
           <Route path="/profile" element={<ProfilePage />} />
