+++ frontend/src/components/navbars/AdminNavBar.tsx
@@ -1,10 +1,12 @@
 import React from "react";
 import { useState } from "react";
 import { useLocation, Link } from "react-router-dom";
-import ProfileButton from "../features/profile/components/ProfileButton.tsx";
-import AddQuestionModal from "../features/questions/components/AddQuestionModal.tsx";
-import { useUser } from "../context/UserContext.tsx";
-import PeerPrepLogo from "./PeerPrepLogo.tsx";
+import ProfileButton from "../../features/profile/components/ProfileButton.tsx";
+import AddQuestionModal from "../../features/questions/components/AddQuestionModal.tsx";
+import { useUser } from "../../context/UserContext.tsx";
+import PeerPrepLogo from "../PeerPrepLogo.tsx";
+import StandardBigButton from "../StandardBigButton.tsx";
+
 interface AdminNavBarProps {
   fetchData: () => Promise<void>;
 }
@@ -25,12 +27,17 @@ const AdminNavBar: React.FC<AdminNavBarProps> = ({ fetchData }) => {
       {location.pathname == "/question" ||
         (location.pathname == "/dashboard" && (
           <div className="container text-off-white">
-            <button
+            {/* <button
               onClick={() => openAddModal()}
               className="bg-green rounded-[25px] p-4 text-2xl hover:bg-emerald-700"
             >
               Add question
-            </button>
+            </button> */}
+            <StandardBigButton
+              onClick={() => openAddModal()}
+              label="Add question"
+              color="green"
+            />
             {isAddModalOpen && (
               <AddQuestionModal fetchData={fetchData} onClose={closeAddModal} />
             )}
