+++ frontend/src/features/dashboard/index.ts
@@ -5,5 +5,8 @@
 export { default as ErrorModal } from "./components/ErrorModals/ErrorModal";
 
 // Export other components directly from 'components'
-export { default as Dashboard } from "./components/Dashboard";
+export { default as DashboardForAdmins } from "./components/DashboardForAdmins";
+export { default as DashboardForUsers } from "./components/DashboardForUsers";
+export { default as DashboardQuestionTable } from "./components/DashboardQuestionTable";
+
 // columns.tsx not for export
