+++ frontend/src/pages/QuestionPage.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import UserNavBar from "../components/UserNavBar.tsx";
+import UserNavBar from "../components/navbars/UserNavBar.tsx";
 import { useParams } from "react-router-dom";
 import { useRetrieveQuestion } from "../features/questions";
 import { QuestionDisplay } from "../features/questions";
